requirements.txt
async-timeout==4.0.3
audioread==3.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.155
botocore==1.34.155
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
crontab==1.0.1
decorator==5.1.1
dropbox==12.0.2
filelock==3.15.4
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-RQ2==18.3
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
freezegun==1.5.1
fsspec==2024.6.1
ftfy==6.2.3
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
httplib2==0.22.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==2.0.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onetimepass==1.0.1
open_clip_torch==2.26.1
openai-clip==1.0.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
ply==3.11
pooch==1.8.2
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rq==1.16.2
rq-scheduler==0.13.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
six==1.16.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.4.0
SQLAlchemy==2.0.32
stone==3.3.1
sympy==1.13.1
threadpoolctl==3.5.0
timm==1.0.8
tokenizers==0.19.1
torch==2.4.0
torchlibrosa==0.1.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
yt-dlp==2024.8.6
